+++ @planetarium/account-web3-secret-storage/src/KeyId.ts
@@ -1,0 +1,3 @@
+export type KeyId = string;
+
+export default KeyId;