+++ @planetarium/account-web3-secret-storage/src/PassphraseEntry.ts
@@ -1,0 +1,8 @@
+import { KeyId } from "./KeyId.js";
+
+export type Passphrase = Uint8Array | string;
+
+export interface PassphraseEntry {
+  authenticate(keyId: KeyId, firstAttempt: boolean): Promise<Passphrase>;
+  configurePassphrase(): Promise<Passphrase>;
+}