+++ @planetarium/account/src/index.ts
@@ -1,5 +1,15 @@
 export { type Account, type ExportableAccount } from "./Account.js";
 export { Address } from "./Address.js";
+export {
+  type AccountDeletion,
+  type AccountGeneration,
+  type AccountImportation,
+  type AccountMetadata,
+  type AccountRetrieval,
+  type ImportableKeyStore,
+  type KeyStore,
+  type MutableKeyStore,
+} from "./KeyStore.js";
 export { type Message } from "./Message.js";
 export { PublicKey, type PublicKeyForm } from "./PublicKey.js";
 export { RawPrivateKey } from "./RawPrivateKey.js";